Added support for error messages returned by DBQuery to functions-event-db.inc.php.

+++ trunk/functions-event-db.inc.php
@@ -3,20 +3,24 @@
 and from the default calendar if the event was submitted to it. */
 function deletefromevent($eventid) {
 	$query = "DELETE FROM vtcal_event WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($eventid)."'";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query); 
+	if (is_string($result)) return $result;
 
 	// delete event from default calendar if it had been forwarded
 	if ( $_SESSION['CALENDAR_ID'] != "default" ) {
 		// delete existing events in default calendar with same id
 		$query = "DELETE FROM vtcal_event WHERE calendarid='default' AND id='".sqlescape($eventid)."'";
-		$result = DBQuery($query ); 
-	} // end: if ( $_SESSION['CALENDAR_ID'] != "default" )
+		$result =& DBQuery($query ); 
+		if (is_string($result)) return $result;
+	}
+	
+	return true;
 }
 
 /* Remove an event from the event_public table (aka: the event will no longer be public) */
 function deletefromevent_public($eventid) {
 	$query = "DELETE FROM vtcal_event_public WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($eventid)."'";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query ); 
 }
 
 /* Remove all repeating entries from the event table (aka: still under review) for the current calendar,
@@ -24,15 +28,18 @@ and from the default calendar if the event was submitted to it. */
 function repeatdeletefromevent($repeatid) {
 	if (!empty($repeatid)) {
 		$query = "DELETE FROM vtcal_event WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND repeatid='".sqlescape($repeatid)."'";
-		$result = DBQuery($query ); 
+		$result =& DBQuery($query); 
+		if (is_string($result)) return $result;
 	
 		// delete event from default calendar if it had been forwarded
 		if ( $_SESSION['CALENDAR_ID'] != "default" ) {
 			// delete existing events in default calendar with same id
 			$query = "DELETE FROM vtcal_event WHERE calendarid='default' AND repeatid='".sqlescape($repeatid)."'";
-			$result = DBQuery($query ); 
-		} // end: if ( $_SESSION['CALENDAR_ID'] != "default" )
+			$result =& DBQuery($query);
+			if (is_string($result)) return $result;
+		}
 	}
+	return true;
 }
 
 /* Remove all repeating entries from the event_public table (aka: the event will no longer be public),
@@ -40,30 +47,35 @@ and from the default calendar if the event was submitted to it. */
 function repeatdeletefromevent_public($repeatid) {
 	if (!empty($repeatid)) {
 		$query = "DELETE FROM vtcal_event_public WHERE calendarid='".$_SESSION['CALENDAR_ID']."' AND repeatid='".sqlescape($repeatid)."'";
-		$result = DBQuery($query ); 
+		$result =& DBQuery($query); 
+		if (is_string($result)) return $result;
 
 		// delete event from default calendar if it had been forwarded
 		if ( $_SESSION['CALENDAR_ID'] != "default" ) {
 			// delete existing events in default calendar with same id
 			$query = "DELETE FROM vtcal_event_public WHERE calendarid='default' AND repeatid='".sqlescape($repeatid)."'";
-			$result = DBQuery($query ); 
-		} // end: if ( $_SESSION['CALENDAR_ID'] != "default" )
+			$result =& DBQuery($query); 
+			if (is_string($result)) return $result;
+		}
 	}
+	return true;
 }
 
 /* Remove all repeating entries from the event table (aka: still under review) for the current calendar. */
 function deletefromrepeat($repeatid) {
 	if (!empty($repeatid)) {
 		$query = "DELETE FROM vtcal_event_repeat WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($repeatid)."'";
-		$result = DBQuery($query ); 
+		$result =& DBQuery($query); 
+		if (is_string($result)) return $result;
 	}
+	return true;
 }
 
-function insertintoevent($eventid,&$event) {
+function insertintoevent($eventid, &$event) {
 	return insertintoeventsql($_SESSION['CALENDAR_ID'],$eventid,$event);
 }
 
-function insertintoeventsql($calendarid,$eventid,&$event) {
+function insertintoeventsql($calendarid, $eventid, &$event) {
 	$changed = date ("Y-m-d H:i:s");
 	$query = "INSERT INTO vtcal_event (calendarid,id,approved,rejectreason,timebegin,timeend,repeatid,sponsorid,displayedsponsor,displayedsponsorurl,title,wholedayevent,categoryid,description,location,price,contact_name,contact_phone,contact_email,url,recordchangedtime,recordchangeduser,showondefaultcal,showincategory) ";
 	$query.= "VALUES ('".sqlescape($calendarid)."','".sqlescape($eventid)."',0,'";
@@ -84,8 +96,11 @@ function insertintoeventsql($calendarid,$eventid,&$event) {
 	$query.= sqlescape($_SESSION["AUTH_USERID"])."','".sqlescape($showondefaultcal)."','";
 	if (isset($event['showincategory'])) { $showincategory = $event['showincategory']; } else { $showincategory = 0; }
 	$query.= sqlescape($showincategory)."')";
-	$result = DBQuery($query ); 
-	echo $result;
+	
+	$result =& DBQuery($query);
+	if (is_string($result)) return $result;
+	
+	//echo $result;
 	return $eventid;
 }
 
@@ -104,10 +119,13 @@ function insertintoevent_public(&$event) {
 	$query.= sqlescape($event['url'])."','".sqlescape($changed)."','";
 	$query.= sqlescape($_SESSION["AUTH_USERID"])."')";
 
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query); 
+	if (is_string($result)) return $result;
+	
+	return true;
 }
 
-function updateevent($eventid,&$event) {
+function updateevent($eventid, &$event) {
 	$changed = date ("Y-m-d H:i:s");
 	$query = "UPDATE vtcal_event SET approved=0, rejectreason='".sqlescape($event['rejectreason']);
 	$query.= "',timebegin='".sqlescape($event['timebegin'])."',timeend='".sqlescape($event['timeend']);
@@ -121,10 +139,13 @@ function updateevent($eventid,&$event) {
 	$query.= "',recordchangedtime='".sqlescape($changed)."',recordchangeduser='".sqlescape($_SESSION["AUTH_USERID"]);
 	$query.= "',showondefaultcal='".sqlescape($event['showondefaultcal'])."',showincategory='".sqlescape($event['showincategory'])."' ";
 	$query.= "WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($eventid)."'";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query );
+	if (is_string($result)) return $result;
+	
+	return true;
 }
 
-function updateevent_public($eventid,&$event) {
+function updateevent_public($eventid, &$event) {
 	$changed = date ("Y-m-d H:i:s");
 	$query = "UPDATE vtcal_event_public SET timebegin='".sqlescape($event['timebegin']);
 	$query.= "',timeend='".sqlescape($event['timeend'])."',repeatid='".sqlescape($event['repeatid']);
@@ -137,10 +158,13 @@ function updateevent_public($eventid,&$event) {
 	$query.= "',url='".sqlescape($event['url'])."',recordchangedtime='".sqlescape($changed);
 	$query.= "',recordchangeduser='".sqlescape($_SESSION["AUTH_USERID"]);
 	$query.= "' WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($eventid)."'";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query);
+	if (is_string($result)) return $result;
+	
+	return true;
 }
 
-function insertintotemplate($template_name,&$event) {
+function insertintotemplate($template_name, &$event) {
 	$changed = date ("Y-m-d H:i:s");
 	$query = "INSERT INTO vtcal_template (calendarid,name,sponsorid,displayedsponsor,displayedsponsorurl,title,wholedayevent,categoryid,description,location,price,contact_name,contact_phone,contact_email,url,recordchangedtime,recordchangeduser) ";
 	$query.= "VALUES ('".sqlescape($_SESSION['CALENDAR_ID'])."','".sqlescape($template_name);
@@ -151,10 +175,13 @@ function insertintotemplate($template_name,&$event) {
 	$query.= "','".sqlescape($event['price'])."','".sqlescape($event['contact_name']);
 	$query.= "','".sqlescape($event['contact_phone'])."','".sqlescape($event['contact_email']);
 	$query.= "','".sqlescape($event['url'])."','".sqlescape($changed)."','".sqlescape($_SESSION["AUTH_USERID"])."')";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query);
+	if (is_string($result)) return $result;
+	
+	return true;
 }
 
-function updatetemplate($templateid,$template_name,&$event) {
+function updatetemplate($templateid, $template_name, &$event) {
 	$changed = date ("Y-m-d H:i:s");
 	$query = "UPDATE vtcal_template SET name='".sqlescape($template_name)."',sponsorid='".sqlescape($event['sponsorid']);
 	$query.= "',displayedsponsor='".sqlescape($event['displayedsponsor'])."',displayedsponsorurl='".sqlescape($event['displayedsponsorurl']);
@@ -165,10 +192,13 @@ function updatetemplate($templateid,$template_name,&$event) {
 	$query.= "',contact_email='".sqlescape($event['contact_email'])."',url='".sqlescape($event['url']);
 	$query.= "',recordchangedtime='".sqlescape($changed)."',recordchangeduser='".sqlescape($_SESSION["AUTH_USERID"]);
 	$query.= "' WHERE sponsorid='".sqlescape($_SESSION["AUTH_SPONSORID"])."' AND id='".sqlescape($templateid)."'";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query);
+	if (is_string($result)) return $result;
+	
+	return true; 
 }
 
-function insertintorepeat($repeatid,&$event,&$repeat) {
+function insertintorepeat($repeatid, &$event, &$repeat) {
 	$repeat['startdate'] = datetime2timestamp($event['timebegin_year'],$event['timebegin_month'],$event['timebegin_day'],0,0,"am");
 	$repeat['enddate'] = datetime2timestamp($event['timeend_year'],$event['timeend_month'],$event['timeend_day'],0,0,"am");
 	$repeatdef = repeatinput2repeatdef($event,$repeat);
@@ -177,13 +207,14 @@ function insertintorepeat($repeatid,&$event,&$repeat) {
 	// write record into repeat table
 	$query = "INSERT INTO vtcal_event_repeat (calendarid,id,repeatdef,startdate,enddate,recordchangedtime,recordchangeduser) ";
 	$query.= "VALUES ('".sqlescape($_SESSION['CALENDAR_ID'])."','".sqlescape($repeatid)."','".sqlescape($repeatdef)."','".sqlescape($repeat['startdate'])."','".sqlescape($repeat['enddate'])."','".sqlescape($changed)."','".sqlescape($_SESSION["AUTH_USERID"])."')";
-	$result = DBQuery($query ); 
-	$repeat['id'] = $repeatid;
+	$result =& DBQuery($query);
+	if (is_string($result)) return $result;
 	
+	$repeat['id'] = $repeatid;
 	return $repeat['id'];
 }
 
-function updaterepeat($repeatid,&$event,&$repeat) {
+function updaterepeat($repeatid, &$event, &$repeat) {
 	$repeat['startdate'] = datetime2timestamp($event['timebegin_year'],$event['timebegin_month'],$event['timebegin_day'],0,0,"am");
 	$repeat['enddate'] = datetime2timestamp($event['timeend_year'],$event['timeend_month'],$event['timeend_day'],0,0,"am");
 	$repeatdef = repeatinput2repeatdef($event,$repeat);
@@ -192,37 +223,46 @@ function updaterepeat($repeatid,&$event,&$repeat) {
 	$query = "UPDATE vtcal_event_repeat SET repeatdef='".sqlescape($repeatdef)."',startdate='";
 	$query.= sqlescape($repeat['startdate'])."',enddate='".sqlescape($repeat['enddate']);
 	$query.= "' WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($repeatid)."'";
-	$result = DBQuery($query ); 
-
+	$result =& DBQuery($query);
+	if (is_string($result)) return $result;
+	
 	return $repeatid;
 }
 
 // Make a non-repeating event public and remove the old event if a previous version existed.
-function publicizeevent($eventid,&$event) {
-	if (!empty($event['repeatid'])) { // if event delivers repeatid that's fine
+function publicizeevent($eventid, &$event) {
+	// if event delivers repeatid that's fine
+	if (!empty($event['repeatid'])) {
 		$r['repeatid'] = $event['repeatid'];
 	}
-	else { // get repeatid from old entry in event_public (important if event changes from recurring to one-time)
-		$result = DBQuery("SELECT repeatid FROM vtcal_event_public WHERE id='".sqlescape($eventid)."'" ); 
+	
+	// get repeatid from old entry in event_public (important if event changes from recurring to one-time)
+	else {
+		$result =& DBQuery("SELECT repeatid FROM vtcal_event_public WHERE id='".sqlescape($eventid)."'" );
+		if (is_string($result)) return $result;
+		
 		if ($result->numRows()>0) { 
-			$r = $result->fetchRow(DB_FETCHMODE_ASSOC,0);
+			$r =& $result->fetchRow(DB_FETCHMODE_ASSOC,0);
 		}
 	}
 
 	if (!empty($r['repeatid'])) { repeatdeletefromevent_public($r['repeatid']); }
 	else { deletefromevent_public($eventid); }
 	
-	$event['id'] = $eventid; // this line should not be necessary but some functions still have a bug that doesn't pass the id in event['id']
+	// this line should not be necessary but some functions still have a bug that doesn't pass the id in event['id']
+	$event['id'] = $eventid;
 	
 	insertintoevent_public($event);
 
-	$result = DBQuery("UPDATE vtcal_event SET approved=1 WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($eventid)."'" ); 
+	$result =& DBQuery("UPDATE vtcal_event SET approved=1 WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND id='".sqlescape($eventid)."'" ); 
+	if (is_string($result)) return $result;
 
 	// forward event to default calendar if that's indicated
 	if ( $_SESSION['CALENDAR_ID'] != "default" ) {
 		// delete existing events in default calendar with same id
 		$query = "DELETE FROM vtcal_event WHERE calendarid='default' AND id='".sqlescape($eventid)."'";
-		$result = DBQuery($query ); 
+		$result =& DBQuery($query );
+		if (is_string($result)) return $result;
 		
 		if ( $event['showondefaultcal'] == 1 ) {
 			// add new event in default calendar (with approved=0)
@@ -233,13 +273,16 @@ function publicizeevent($eventid,&$event) {
 		} 
 		else {
 			$query = "DELETE FROM vtcal_event_public WHERE calendarid='default' AND id='".sqlescape($eventid)."'";
-			$result = DBQuery($query ); 
+			$result =& DBQuery($query ); 
+			if (is_string($result)) return $result;
 		}
-	} // end: if ( $_SESSION['CALENDAR_ID'] != "default" )
+	}
+	
+	return true;
 } // end: publicizeevent
 
 // Make a repeating event public and remove the old event if a previous version existed.
-function repeatpublicizeevent($eventid,&$event) {
+function repeatpublicizeevent($eventid, &$event) {
 	deletefromevent_public($eventid);
 	if (!empty($event['repeatid'])) {
 		repeatdeletefromevent_public($event['repeatid']);
@@ -254,27 +297,35 @@ function repeatpublicizeevent($eventid,&$event) {
 			$e = substr($e,0,$dashpos); 
 		}
 		$query = "DELETE FROM vtcal_event WHERE calendarid='default' AND id='".sqlescape($e)."'";
-		$result = DBQuery($query ); 
+		$result =& DBQuery($query ); 
+		if (is_string($result)) return $result;
 		
 		if (!empty($event['repeatid'])) {
 			$query = "DELETE FROM vtcal_event WHERE calendarid='default' AND repeatid='".sqlescape($event['repeatid'])."'";
-			$result = DBQuery($query ); 
+			$result =& DBQuery($query ); 
+			if (is_string($result)) return $result;
 		}
 		
-		if ( $event['showondefaultcal'] != 1 ) { // remove events if checkmark for forwarding is removed
+		// remove events if checkmark for forwarding is removed
+		if ( $event['showondefaultcal'] != 1 ) {
 			$query = "DELETE FROM vtcal_event_public WHERE calendarid='default' AND id='".sqlescape($e)."'";
-			$result = DBQuery($query ); 
+			$result =& DBQuery($query ); 
+			if (is_string($result)) return $result;
+			
 			if (!empty($event['repeatid'])) {
 				$query = "DELETE FROM vtcal_event_public WHERE calendarid='default' AND repeatid='".sqlescape($event['repeatid'])."'";
-				$result = DBQuery($query ); 
+				$result =& DBQuery($query ); 
+				if (is_string($result)) return $result;
 			}
 		}
-	} // end: if ( $_SESSION['CALENDAR_ID'] != "default" )
+	}
 
 	// copy all events into event_public
-	$result = DBQuery("SELECT * FROM vtcal_event WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND repeatid='".sqlescape($event['repeatid'])."'" );
+	$result =& DBQuery("SELECT * FROM vtcal_event WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND repeatid='".sqlescape($event['repeatid'])."'" );
+	if (is_string($result)) return $result;
+	
 	for ($i=0;$i<$result->numRows();$i++) {
-		$event = $result->fetchRow(DB_FETCHMODE_ASSOC,$i);
+		$event =& $result->fetchRow(DB_FETCHMODE_ASSOC,$i);
 //    eventaddslashes($event);
 		insertintoevent_public($event);
 		
@@ -287,10 +338,13 @@ function repeatpublicizeevent($eventid,&$event) {
 				insertintoeventsql("default",$event['id'],$event);
 				$event['categoryid'] = $eventcategoryid;
 			}
-		} // end: if ( $_SESSION['CALENDAR_ID'] != "default" )
-	} // end: for(...
+		}
+	}
 
 	$query = "UPDATE vtcal_event SET approved=1 WHERE calendarid='".sqlescape($_SESSION['CALENDAR_ID'])."' AND approved=0 AND repeatid='".sqlescape($event['repeatid'])."'";
-	$result = DBQuery($query ); 
+	$result =& DBQuery($query ); 
+	if (is_string($result)) return $result;
+	
+	return true;
 } // end: repeatpublicizeevent
 ?>
